requirements.txt
django
django_extensions
djangorestframework
markdown
django-filter